Improves `TCPaginationHelpers` implementation
